UI folder hierarchy refactor to group files per component
