+++ webui/src/app/components/client/AlignmentEntry/types.ts
@@ -1,0 +1,6 @@
+export interface geneInfo {
+    readonly id: string,
+    readonly symbol: string,
+    readonly species: any,
+    readonly genomeLocations: Array<any>
+}
