+++ webui/src/app/components/client/JobSubmitForm/serverActions.ts
@@ -2,7 +2,7 @@
 
 import { env } from 'process'
 
-import { jobType, geneInfo } from "./types";
+import { jobType } from "./types";
 
 const API_BASE = (env.PAVI_API_BASE_URL || 'http://localhost:8000')+'/api'
 
@@ -91,47 +91,4 @@ export async function submitNewPipelineJob (inputStr: string): Promise<jobType> 
             inputValidationPassed: false
         }
     }
-}
-
-export async function fetchGeneInfo (geneId: string): Promise<geneInfo|undefined> {
-
-    console.log(`New gene info request received.`)
-
-    const jobResponse = fetch(`https://www.alliancegenome.org/api/gene/${geneId}`, {
-        method: 'GET',
-        headers: {
-            'accept': 'application/json'
-        }
-    })
-    .then((response: Response) => {
-        if ( 500 <= response.status && response.status <= 599 ){
-            // No point in attempting to process the body, as no body is expected.
-            throw new Error('Server error received.', {cause: 'server error'})
-        }
-
-        return Promise.all([Promise.resolve(response), response.json()]);
-    })
-    .then(([response, body]) => {
-        if (response.ok) {
-            console.log(`Gene info for gene ${geneId} received successfully: ${JSON.stringify(body)}`)
-            return body as geneInfo;
-        } else {
-            const errMsg = 'Failure response received from gene API.'
-            console.error(errMsg)
-            if( 400 <= response.status && response.status <= 499 ){
-                throw new Error(errMsg, {cause: 'user error'})
-            }
-            else{
-                console.log('Non user-error response received:', response)
-                throw new Error(errMsg, {cause: 'unkown'})
-            }
-
-        }
-    })
-    .catch((e: Error) => {
-        console.error('Error caught while requesting gene info:', e)
-        return undefined;
-    });
-
-    return jobResponse
 }
