+++ webui/src/app/components/client/JobSubmitForm/types.ts
@@ -4,22 +4,6 @@ export interface jobType {
     inputValidationPassed?: boolean
 }
 
-export interface geneInfo {
-    readonly id: string,
-    readonly symbol: string,
-    readonly species: any,
-    readonly genomeLocations: Array<any>
-}
-
-export interface transcriptInfo {
-    readonly id: string,
-    readonly name: string,
-    readonly exons: Array<{
-        refStart: number
-        refEnd: number
-    }>
-}
-
 export interface jobSumbissionPayloadRecord {
     name: string,
     seq_id: string,
